dom/html/test/test_bug109445.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=109445
-->
<head>
  <title>Test for Bug 109445</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=109445">Mozilla Bug 109445</a>
<p id="display">
<map name=a>
<area shape=rect coords=25,25,75,75 href=#x>
</map>
<map id=b>
<area shape=rect coords=25,25,75,75 href=#y>
</map>
<map name=a>
<area shape=rect coords=25,25,75,75 href=#FAIL>
</map>
<map id=b>
<area shape=rect coords=25,25,75,75 href=#FAIL>
</map>

<img usemap=#a src=image.png>
<img usemap=#b src=image.png>
</p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 109445 **/
SimpleTest.waitForExplicitFinish();
var images = document.getElementsByTagName("img");
var second = false;
onhashchange = function() {
  if (!second) {
    second = true;
    is(location.hash, "#x", "First map");
    SimpleTest.waitForFocus(function() synthesizeMouse(images[1], 50, 50, {}));
  } else {
    is(location.hash, "#y", "Second map");
    SimpleTest.finish();
  }
};
SimpleTest.waitForFocus(function() synthesizeMouse(images[0], 50, 50, {}));
</script>
</pre>
</body>
</html>
